Make docker get_latest_image support datestamp tags
The docker `engine.get_latest_image_id_for_service` was hardcoded to
always return '[project]:latest' images, but when running a service, the
builder tags containers with the date, like
`mk-ii-redis:20170228152303`. To handle this, when `:latest` does not
exist, the function will list all the images with the given name and
return the one with the newest datetime tag.